Split sceNp into sceNpUtil and register sceNpClans
